Update similarity measure to match SimPolyProtect implementation

+++ src/main/java/com/simprints/simface/matcher/MatchProcessor.kt
@@ -1,9 +1,9 @@
 package com.simprints.simface.matcher
 
 interface MatchProcessor {
-    fun verificationScore(probe: ByteArray, matchAgainst: ByteArray): Float
+    fun verificationScore(probe: ByteArray, matchAgainst: ByteArray): Double
     fun identificationScore(
         probe: ByteArray,
         matchReferences: List<ByteArray>
-    ): Map<ByteArray, Float>
+    ): Map<ByteArray, Double>
 }
